+++ src/lib/st-breadcrumbs/st-breadcrumbs.component.spec.ts
@@ -8,12 +8,11 @@
  *
  * SPDX-License-Identifier: Apache-2.0.
  */
-import { SimpleChange } from '@angular/core';
+import { ChangeDetectionStrategy, NO_ERRORS_SCHEMA, SimpleChange } from '@angular/core';
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
-import { NO_ERRORS_SCHEMA } from '@angular/core';
 
 import { StBreadCrumbsComponent } from './st-breadcrumbs.component';
-import { StBreadCrumbItem } from './st-breadcrumbs.interface';
+import { StBreadCrumbItem, StBreadCrumbMode } from './st-breadcrumbs.interface';
 
 const item: StBreadCrumbItem = {
    id: 'home',
@@ -29,7 +28,7 @@ describe('StBreadCrumbsComponent', () => {
       { id: 'home', icon: 'icon-home' },
       { id: 'section1', label: 'section1', icon: '' },
       { id: 'section2', label: 'section2', icon: '' },
-      { id: 'section3', label: 'section3'},
+      { id: 'section3', label: 'section3' },
       { id: 'section4', label: 'section4', icon: 'icon-home' },
       { id: 'section5', label: 'section5', icon: '' },
       { id: 'section6', label: 'section6', icon: '' }];
@@ -41,7 +40,11 @@ describe('StBreadCrumbsComponent', () => {
          TestBed.configureTestingModule({
             declarations: [StBreadCrumbsComponent],
             schemas: [NO_ERRORS_SCHEMA]
-         }).compileComponents(); // compile template and css
+         })
+            .overrideComponent(StBreadCrumbsComponent, {
+               set: { changeDetection: ChangeDetectionStrategy.Default }
+            })
+            .compileComponents(); // compile template and css
       })
    );
 
@@ -62,7 +65,7 @@ describe('StBreadCrumbsComponent', () => {
       });
 
       it('should be initialized undefined', () => {
-         const itemUndefined: StBreadCrumbItem = {id: ''};
+         const itemUndefined: StBreadCrumbItem = { id: '' };
          component.options[1] = itemUndefined;
          fixture.detectChanges();
          expect(component.hasLabel(1)).toBeFalsy();
@@ -133,6 +136,7 @@ describe('StBreadCrumbsComponent', () => {
       fixture.detectChanges();
       expect(component.getIcon(4)).toEqual(String(menuMock[4].icon));
    });
+
    it('Should get 3 dots when index is -1', () => {
       fixture.detectChanges();
       expect(component.getLabel(-1)).toEqual('...');
@@ -160,8 +164,28 @@ describe('StBreadCrumbsComponent', () => {
 
       expect(component.indexArray).toEqual([0, 1, 2, 3, 4, 5, 6]);
 
-      component.options = [...menuMock, {id: 'section7', label: 'section7'}];
-      component.ngOnChanges({ options: new SimpleChange(menuMock, component.options, true)});
+      component.options = [...menuMock, { id: 'section7', label: 'section7' }];
+      component.ngOnChanges({ options: new SimpleChange(menuMock, component.options, true) });
       expect(component.indexArray).toEqual([0, -1, 2, 3, 4, 5, 6, 7]);
+   });
+
+   describe('Mode can be configured in order to modify its design', () => {
+      it('If mode is not introduced as input, it will be DEFAULT', () => {
+         fixture.detectChanges();
+
+         expect(fixture.nativeElement.querySelector('ul').classList).toContain('default-mode');
+      });
+
+      it('If mode is introduced as input, it will be added as class to the host', () => {
+         component.mode = StBreadCrumbMode.TITLE;
+         fixture.detectChanges();
+
+         expect(fixture.nativeElement.querySelector('ul').classList).toContain('title-mode');
+
+         component.mode = StBreadCrumbMode.DEFAULT;
+         fixture.detectChanges();
+
+         expect(fixture.nativeElement.querySelector('ul').classList).toContain('default-mode');
+      });
    });
 });