+++ src/lib/st-breadcrumbs/st-breadcrumbs.interface.ts
@@ -14,3 +14,5 @@ export class StBreadCrumbItem {
    label?: string;
    icon?: string;
 }
+
+export enum StBreadCrumbMode { DEFAULT =  <any> 'default', TITLE =  <any> 'title'}